pyproject.toml
[project]
name = "autoreg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncio>=4.0.0",
    "langchain>=0.3.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "langchain-milvus>=0.3.3",
    "pypdf>=6.6.2",
    "python-dotenv>=1.2.1",
    "langchain-ollama>=1.0.0",
    "langchain-core>=1.0.7",
    "langchain[google-genai]>=1.0.8",
    "rich>=14.3.1",
    "pymilvus>=2.6.7",
    "openai>=2.15.0",
    "google-generativeai>=0.8.6",
    "pypdf2>=3.0.1",
    "langchain-text-splitters>=1.1.0",
    "pydantic>=2.12.5",
    "spelling>=0.1.8",
    "datasets>=4.5.0",
    "torch>=2.10.0",
    "transformers<4.45",
    "sentence-transformers>=5.2.2",
    "peft>=0.18.1",
    "flagembedding>=1.3.5",
    "pdfplumber>=0.11.9",
]

[tool.ruff]
line-length = 88
target-version = "py313"
fix = true
show-fixes = true
indent-width = 4
exclude = [".venv"]




[[tool.uv.index]]
name = "a5x-pypi"
url = "https://nexus.prod.automation.devops.a5x.com.br/repository/pypi-a5x/simple"
